Extent Utilities

Introduction

These are a few methods to generate and work with Extents.Extent objects. Most of this will ideally be moved to GeoInterface.jl or to Extents.jl itself. Until then, we keep it here.

API

CoolWalksUtils.apply_extent!

CoolWalksUtils.extent_center

CoolWalksUtils.extent_contains

CoolWalksUtils.extent_contains

CoolWalksUtils.extent_contains

CoolWalksUtils.geoiter_extent

CoolWalksUtils.geoiter_extent

CoolWalksUtils.apply_extent! — Method

apply_extent!(df, extent; source=[:lon, :lat])

trim df to only contain rows where extent_contains(extent, source...) is true.

CoolWalksUtils.extent_center — Method

extent_center(extent)

calculates the centerpoint of extent. Return a NamedTuple with the same keys as extent.

CoolWalksUtils.extent_contains — Method

extent_contains(a::Extent, X, Y)

Checks if the extent a contains the point given by X and Y.

CoolWalksUtils.extent_contains — Method

extent_contains(a::Extent, geometry)

Checks if the GeoInterface.extent of geometry is fully contained in extent a.

CoolWalksUtils.extent_contains — Method

extent_contains(a::Extent, b::Extent)

Checks if the extent b is fully contained in extent a.

CoolWalksUtils.geoiter_extent — Method

geoiter_extent(X, Y)

constructs the Extent from two iterables representing x and y coordinates. (Note that Numbers are iterable: geoiter_extent(3.5, 3) works.)

CoolWalksUtils.geoiter_extent — Method

geoiter_extent(geo_iter)

constructs the Extent from an iterable with GeoInterface compatible elements.