Maths

Introduction

These are just a few pure maths functions used all over the place.

API

CoolWalksUtils.cross

CoolWalksUtils.is_left

CoolWalksUtils.unit

CoolWalksUtils.cross — Method

cross(v1, v2)

cross product of the two input vectors. Stolen from LinearAlgebra.jl.

CoolWalksUtils.is_left — Method

is_left(center_point, test_point, sunpos::AbstractArray)
is_left(center_point::AbstractArray, test_point::AbstractArray, sunpos::AbstractArray)

checks if test_point is left of the line going through center_point in the direction of -sunpos. (sunpos points AT the sun.)

CoolWalksUtils.unit — Method

unit(vector)

returns the vector scaled to unity (using L2 norm).