Projection utilities

Introduction

CoolWalkUtils.jl specifies a few convenient functions to interpret and reproject different datastructures revolving around ArchGDAL.jl types. (most of this behaviour should probably be part of ArchGDAL.jl, but it is not, as of the time of writing.)

API

CoolWalksUtils.OSM_ref

CoolWalksUtils._execute_projection!

CoolWalksUtils.apply_wsg_84!

CoolWalksUtils.crs_local

CoolWalksUtils.in_local_coordinates

CoolWalksUtils.project_back!

CoolWalksUtils.project_local!

CoolWalksUtils.reinterp_crs!

CoolWalksUtils.OSM_ref — Constant

Reference which holds the WSG84 (EPSG4326) ArchGDAL Spatial Reference System with lon-lat order.

CoolWalksUtils._execute_projection! — Method

_execute_projection!(geometry::ArchGDAL.IGeometry, src, dst)
_execute_projection!(geometry_iterable, src, dst)

Applies the transformation from src to dst to geometry or the elements of geometry_iterable.

CoolWalksUtils.apply_wsg_84! — Method

apply_wsg_84!(geometry_container)

reinterprets the coordinates of geometry_container (either ArchGDAL.IGeometry or some iterable of the same) to be in WSG84.

CoolWalksUtils.crs_local — Method

crs_local(lon, lat)

builds Transverse Mercator ArchGDAL coordinate reference system centered around lon and lat.

CoolWalksUtils.in_local_coordinates — Method

in_local_coordinates(f, obs::ShadowObservatory, args...)
in_local_coordinates(f, df::DataFrame, args...)

projects args... to local crs determined by obs, passes all args... to f, projects args... back to global crs and returns the return value of f.

If a DataFrame is passed as the second entry, the observatory for local projection is taken from metadata(df, "observatory").

CoolWalksUtils.project_back! — Function

project_back!(geometry_container)

Projects geometry or containers of geometry (iterables) from their respective local system back to WSG84.

In the case of iterables, we assume all the entries to be in the same crs as the first entry. For DataFrames each colum containing geometry (isgeometry(first(column))) is assumed to be in the same crs as the first entry.

Returns the passed in geometry container.

CoolWalksUtils.project_local! — Function

project_local!(geometry_container, observatory::ShadowObservatory)
project_local!(df::DataFrame, observatory::ShadowObservatory=metadata(df, "observatory"))

Projects geometry or containers of geometry (iterables) to transverse mercator centered at (lon, lat) in observatory.

In the case of iterables, we assume all the entries to be in the same crs as the first entry. For DataFrames each colum containing geometry (determined by isgeometry(first(column))) is assumed to be in the same crs as the first entry.

Returns the passed geometry container.

project_local!(container, lon, lat)

Same as above, but allows to specify the center of the projection via lon and lat in degrees.

CoolWalksUtils.reinterp_crs! — Method

reinterp_crs!(geometry_container, crs)

reinterprets the coordinates of geometry_container (either ArchGDAL.IGeometry or some iterable of the same) to be in crs.