Maths

Introduction

These are just a few, geometry related maths functions used everywhere.

API

CoolWalksUtils.intersection_distances

CoolWalksUtils.is_ccw

CoolWalksUtils.is_convex

CoolWalksUtils.is_left

CoolWalksUtils.switches_side

CoolWalksUtils.intersection_distances — Method

intersection_distances(a, b, c, d)

returns the solution to a * (1-x1) + b * x1 == c * (1-x2) + d * x2.

julia> intersection_distances([0,0], [1,1], [5,0], [0,5])
2-element Vector{Float64}:
 2.5
 0.5

julia> intersection_distances([0,0.0], [1,1], [4.0, 0.0], [10.0, 2.0])
2-element Vector{Float64}:
 -2.0
 -1.0

CoolWalksUtils.is_ccw — Method

is_ccw(a, b, c)

tests if the 2D points a, b, c form a counterclockwise triangle.

julia> is_ccw([0,0], [1,0], [1,1])
true

julia> is_ccw([0,0], [0,1], [1.4, 0.5])
false

CoolWalksUtils.is_convex — Method

is_convex(polygon)

tests if the exterior of the GeoInterface compatible polygon is convex.

is_convex(points::AbstractVector)

tests if vector of 2D points points describes a convex polygon.

Assumes that points[1] == points[end].

julia> square = [[0, 0], [1, 0], [1, 1], [0, 1], [0, 0]];

julia> is_convex(square)
true

julia> thing = [[0.6, 0.6], [1.0, 0.0], [1.0, 1.0], [0.0, 1.0], [0.6, 0.6]];

julia> is_convex(thing)
false

CoolWalksUtils.is_left — Method

checks if test_point is left of the line going through center_point in the direction of -sunpos. (sunpos points AT the sun, has three entries. The last on will be set to 0.)

Signatures

is_left(center_point, test_point, sunpos::AbstractVector)

center_point and test_point can be any GeoInterface compatible points.

is_left(center_point::AbstractVector, test_point::AbstractVector, sunpos::AbstractVector)

center_point and test_point can be eiter 3 or 2 entry long vectors. In the later case, we add a third entry ==0.

Examples

julia> is_left([0, 0], [1, 1], [-1, 0, 0])
true

julia> is_left([0.5, 0.5, 0.0], [1, 1, 0.0], [1, 0, 0])
false

CoolWalksUtils.switches_side — Method

switches_side(a, b, c, d)

tests if the line defined by the 2D points a, b intersects the line segment defined by the 2D points c,d.

julia> switches_side([0,0], [1,1], [5,0], [0,5])
true

julia> switches_side([0,0], [1,1], [0.4, 0.2], [10.5, 2.0])
false