Projection utilities

Introduction

CoolWalkUtils.jl specifies a few convenient functions to interpret and reproject different datastructures revolving around ArchGDAL.jl types. (most of this behaviour should probably be part of ArchGDAL.jl, but it is not, as of the time of writing.)

API

CoolWalksUtils.OSM_ref

CoolWalksUtils.apply_wsg_84!

CoolWalksUtils.project_back!

CoolWalksUtils.project_geo_array!

CoolWalksUtils.project_local!

CoolWalksUtils.reinterp_crs!

CoolWalksUtils.OSM_ref — Constant

Reference which holds the WSG84 (EPSG4326) ArchGDAL Spatial Reference System with lon-lat order.

CoolWalksUtils.apply_wsg_84! — Method

apply_wsg_84!(geom)

reinterprets the coordinates of ArchGDAL.jl geometry geom to be in OSM_ref (EPSG4326).

CoolWalksUtils.project_back! — Method

project_back!(geo_array)

projects iterable of ArchGDAL.jl geometries from the coordinate system of first(geo_array) to the coordinate reference system given in OSM_ref (EPSG4326).

CoolWalksUtils.project_geo_array! — Method

project_geo_array!(geo_array, trans)

applies the ArchGDAL transformation to every element in geo_array.

CoolWalksUtils.project_local! — Method

project_local!(geo_array, center_lon, center_lat)

projects iterable of ArchGDAL.jl geometries from the coordinate system of first(geo_array) to the transverse mercator projection centered at center_lon, center_lat.

CoolWalksUtils.reinterp_crs! — Method

reinterp_crs!(geom, crs)

reinterprets the coordinates of ArchGDAL.jl geometry geom to be in the coordinate reference system crs.