RTree Building

Introduction

Just two little functions to build RTrees out of ArchGDAL polygons or ArchGDAL points. And another function to get the bounding box with a buffer of any geometry as a SpatialIndexing.Rect.

API

CoolWalksUtils.build_point_rtree

CoolWalksUtils.build_rtree

CoolWalksUtils.build_rtree

CoolWalksUtils.rect_from_geom

CoolWalksUtils.build_point_rtree — Function

build_point_rtree(points, data, include_orig_point=false)

builds SpatialIndexing.RTree{Float64, 2} from an array containing ArchGDAL.wkbPoints. If include_orig_point is false, the val of an entry in the tree is just the corresponding element in the data array. If include_orig_point is true, it will be a NamedTuple with keys of :data and :pointgeom.

CoolWalksUtils.build_rtree — Method

build_rtree(geo_arr)

builds SpatialIndexing.RTree{Float64, 2} from an array containing ArchGDAL.wkbPolygons. The value of an entry in the RTree is a named tuple with: (orig=original_geometry,prep=prepared_geometry). orig is just the original object, an element from geo_arr, where prep is the prepared geometry, derived from orig. The latter one can be used in a few ArchGDAL functions to get higher performance, for example in intersection testing, because relevant values get precomputed and cashed in the prepared geometry, rather than precomputed on every test.

CoolWalksUtils.build_rtree — Method

build_rtree(df::DataFrame)

builds SpatialIndexing.RTree{Float64, 2} from a DataFrame containing at least a column named geometry. The value of an entry in the RTree is a named tuple with: (prep=prepared_geometry, row=dataframe_row). prep is the prepared geometry, derived from row.geometry and can be used in a few ArchGDAL functions to get higher performance, for example in intersection testing, because relevant values get precomputed and cashed in the prepared geometry, rather than precomputed on every test. The row entry is a reference to the row of the original dataframe, providing access to all relevant data.

CoolWalksUtils.rect_from_geom — Method

rect_from_geom(geom; buffer=0.0)

builds SpatialIndexing.Rect with the extent of the geometry geom and a buffer of buffer added to each edge of the Rect. This is for example useful when we want to query around points, since them having a zero area extend might mess with numerical stuff.