Sun Position

Introduction

Roberto Grena,
Five new algorithms for the computation of sun position from 2010 to 2110,
Solar Energy,
Volume 86, Issue 5,
2012,
Pages 1323-1337,
ISSN 0038-092X,
https://doi.org/10.1016/j.solener.2012.01.024.
(https://www.sciencedirect.com/science/article/pii/S0038092X12000400)

API

CoolWalksUtils.algorithm_1

CoolWalksUtils.get_local_sun_pos

CoolWalksUtils.sunposition

CoolWalksUtils.algorithm_1 — Method

algorithm_1(t_2060, longitude)

algorithm 1 to calculate the sun position as described by Roberto Grena (2012), Five new algorithms for the computation of sun position from 2010 to 2110, Solar Energy, 86(5):1323–1337, doi:10.1016/j.solener.2012.01.024. Simplified to not use TT-UT, pressure and temperature.

CoolWalksUtils.get_local_sun_pos — Method

get_local_sun_pos(latitude, declination, hour_angle)

transform the location of the sun in the sky from global coordinates into spherical coordinates local to latitude and longitude used to calculate the inputs.

Returns

(elevation, azimuth) The elevation is the heigth of the sun above the horizon, in radians. The azimuth is the horizontal position of the sun in a polar coordinate system, where azimuth=0 points to the south pole. negative angles point to the east, positive to the west. again, in radians.

CoolWalksUtils.sunposition — Function

Functions to calculate the sun position, adapted from Roberto Grena (2012), Five new algorithms for the computation of sun position from 2010 to 2110, Solar Energy, 86(5):1323–1337, doi:10.1016/j.solener.2012.01.024.

we use only algorithm 1 and omit the inputs of TT-UT, pressure and temperature, since we do not need the precision to launch satellites

Notes on Arguments

longitude has to be in radians from 0 to 2π, starting from Greenwitch going east

latitude has to be in radians from -π/2 to -π/2, starting from south pole, going north

Return

Array with 3 entries [x,y,z], representing vector pointing towards sun in local, cartesian coordinate system, centered at longitude, latitude with x pointing east, y pointing north and z pointing up.

sunposition(date::DateTime, longitude, latitude, timezone::Int=1, daylight_saving::Bool=true)

Notes on Arguments

date is assumed to be in local time.

timezone and daylight_saving are used to convert to gmt. If the time is allready gmt, set timezone=0 and daylight_saving=false

sunposition(time, day::Int, month::Int, year::Int, longitude, latitude, timezone::Int=1, daylight_saving::Bool=true)

Notes on Arguments

time is local time in hours. Convert minutes and seconds... to fractions of hours. (13:30 becomes 13.5) range from 0 to 24.

timezone and daylight_saving are used to convert to gmt. If the time is allready gmt, set timezone=0 and daylight_saving=false.